# scan-coordinator/scan-coordinator-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SentinelFlow Security Platform",
    "description": "Unified security scan platform: trigger SAST + SCA scans, query vulnerabilities and BOM from Black Duck, and request AI-powered auto-fix via Claude Code.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://arborescent-actorly-carroll.ngrok-free.dev",
      "description": "SentinelFlow (ngrok)"
    }
  ],
  "paths": {
...
    "/fix/prepare": {
      "post": {
        "tags": [
          "Fix"
        ],
        "summary": "Filter high-risk vulnerabilities and prepare fix plan for Bob",
        "description": "從掃描報告中過濾 HIGH / CRITICAL 漏洞，整理成 Bob 可直接讀取的修復計畫 JSON。\n\n- 支援 scan_id = \"latest\" 自動取最新報告\n- SCA：依套件分組，列出 CVE 清單\n- SAST：依問題類型分組，附修復指引\n- auto_submit=true：自動送出給 Bob 執行修復",
        "operationId": "prepareFix",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FixPrepareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/fix/submit": {
      "post": {
        "tags": [
          "Fix"
        ],
        "summary": "Prepare fix plan and automatically submit to Bob for execution",
        "description": "等同於 /fix/prepare 並強制 auto_submit=true。\n一個指令完成：過濾高風險 → 整理 → 送給 Bob 修復。",
        "operationId": "submitFix",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FixPrepareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/register": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "手動登記專案（路徑不在 PROJECTS_ROOT 下時使用）",
        "description": "手動登記一個專案，覆寫自動發現的設定。\n適合專案不在 PROJECTS_ROOT 下、或 Black Duck 專案名稱不符合預設規則的情況。",
        "operationId": "registerProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/resolve/{project_name}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "查詢專案名稱對應的路徑",
        "description": "查詢專案名稱對應的完整路徑與 Black Duck 專案設定。",
        "operationId": "resolveProject",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FixPrepareRequest": {
        "properties": {
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          },
          "auto_submit": {
            "title": "Auto Submit",
            "default": false,
            "type": "boolean"
          },
          "project_root": {
            "title": "Project Root",
            "type": "string"
          },
          "project": {
            "title": "Project",
            "type": "string"
          }
        },
        "type": "object",
        "required": [
          "scan_id"
        ],
        "title": "FixPrepareRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProjectRegisterRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "bd_project": {
            "title": "Bd Project",
            "type": "string"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectRegisterRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}